refactor: move compress-commons to zip-stream
