--- packages/compress-commons/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@archiver/compress-commons",
-  "version": "0.0.7",
-  "description": "a library that defines a common interface for working with archive formats within node",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/node-archiver/archiver.git",
-    "directory": "packages/compress-commons"
-  },
-  "files": [
-    "dist"
-  ],
-  "type": "module",
-  "exports": {
-    ".": "./dist/index.mjs",
-    "./constants": "./dist/constants.mjs",
-    "./utils": "./dist/utils.mjs",
-    "./package.json": "./package.json"
-  },
-  "scripts": {
-    "build": "tsdown",
-    "test": "bun test --reporter dot",
-    "prepublishOnly": "bun run build"
-  },
-  "devDependencies": {
-    "tsdown": "catalog:"
-  },
-  "engines": {
-    "node": ">=24"
-  }
-}
