--- packages/compress-commons/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "rootDir": "./"
-  },
-  "include": ["src", "tests"]
-}
