--- packages/compress-commons/tsdown.config.ts
@@ -1,7 +1,0 @@
-import { defineConfig } from "tsdown";
-
-export default defineConfig({
-  entry: ["src/index.ts", "src/constants.ts", "src/utils.ts"],
-  dts: true,
-  exports: true,
-});
