+++ packages/zip-stream/package.json
@@ -14,6 +14,9 @@
   "type": "module",
   "exports": {
     ".": "./dist/index.mjs",
+    "./compress-commons": "./dist/compress-commons/index.mjs",
+    "./compress-commons/constants": "./dist/compress-commons/constants.mjs",
+    "./compress-commons/utils": "./dist/compress-commons/utils.mjs",
     "./utils": "./dist/utils.mjs",
     "./package.json": "./package.json"
   },
