+++ packages/zip-stream/src/compress-commons/archivers/archive-entry.ts
