+++ packages/zip-stream/src/compress-commons/archivers/archive-output-stream.ts
