+++ packages/zip-stream/src/compress-commons/archivers/crc32-stream.ts
