+++ packages/zip-stream/src/compress-commons/archivers/general-purpose-bit.ts
