+++ packages/zip-stream/src/compress-commons/archivers/unix-stat.ts
