+++ packages/zip-stream/src/compress-commons/archivers/zip-archive-entry.ts
