+++ packages/zip-stream/src/compress-commons/constants.ts
