+++ packages/zip-stream/src/compress-commons/utils.ts
