+++ packages/zip-stream/tests/compress-commons/fixtures/directory/level0.txt
