+++ packages/zip-stream/tests/compress-commons/fixtures/directory/subdir/level1.txt
