+++ packages/zip-stream/tests/compress-commons/fixtures/directory/subdir/subsub/level2.txt
