+++ packages/zip-stream/tests/compress-commons/fixtures/image.png
