+++ packages/zip-stream/tests/compress-commons/fixtures/test.txt
