+++ packages/zip-stream/tsdown.config.ts
@@ -1,7 +1,13 @@
 import { defineConfig } from "tsdown";
 
 export default defineConfig({
-  entry: ["src/index.ts", "src/utils.ts"],
+  entry: [
+    "src/index.ts",
+    "src/utils.ts",
+    "src/compress-commons/index.ts",
+    "src/compress-commons/constants.ts",
+    "src/compress-commons/utils.ts",
+  ],
   dts: true,
   exports: true,
 });
